Split out getPostsLastPageForQuery
